Detect inline link reference titles via regex capture (#292)
Extend LINK_REF_RE with an optional title capture group so
link_ref_needs_title no longer misclassifies destinations ending in ')'.
Move link reference wrap tests into a dedicated submodule.
